Flytter temp tester og utils til egne filer
