+++ src/main/kotlin/no/nav/tilleggsstonader/sak/util/TakeIfUtil.kt
@@ -1,0 +1,7 @@
+package no.nav.tilleggsstonader.sak.util
+
+object TakeIfUtil {
+
+    inline fun <reified T> Any.takeIfType(): T? =
+        takeIf { it is T }?.let { it as T }
+}
